make function global in user service
